prevent tooltip text from overflowing background
Previously, if the replacement text was shorter than the original text,
this resulted in the text overflowing the black background. Now we use
word-wrap to avoid that ugly mess.

This does introduce a lesser issue in that the prices don't typically
break in appealing manners, but at least they're legible.

closes #2